domain_to_str(): PyString_FromStringAndSize -> PyString_FromString
this fixes a regression introduced in d04d3919. wdns_domain_to_str()
does not return the number of bytes written to dst, it returns the
number of bytes read from src. dst is guaranteed to be NUL-terminated
after a call to wdns_domain_to_str(), so switch back to using
PyString_FromString() to convert dst to a python string.