Encode to and from US-ASCII to avoid overusing the 'bytes' type
The DNS presentation format uses the US-ASCII character set de facto, so
it ought to be safe to require presentation format strings to be strings
that happen to only use characters that are present in US-ASCII. For
instance, it should be possible to write:

    wdns.str_to_rrtype('A')

Rather than:

    wdns.str_to_rrtype(b'A')

This commit alters the Python wrappers for domain_to_str(),
opcode_to_str(), rcode_to_str(), rrclass_to_str(), rrtype_to_str(), and
rdata_to_str() to call .decode('ascii') on the byte array that we get
back from libwdns. This should never fail because libwdns should always
hand us back a US-ASCII string.

This commit also alters the Python wrappers for str_to_rrtype() and
str_to_name() to accept Python strings that use the US-ASCII subset of
characters by calling .encode('ascii') on the string that we get from
the Python caller. This can fail if the caller passes a string that
can't be encoded into US-ASCII, but this should properly fail -- if the
caller needs to pass non-ASCII data in the presentation form of a DNS
name, it should be escaped with ASCII-compatible DNS-style \DDD escape
literals.