Fix rdata_to_str() to return type 'str' on Python 2
Calling .decode('ascii') is not quite what we need for Python 2, because
this changes the return type to 'unicode'. Wrap the value in a str() so
that we get the right return type.

The underlying data returned by the C wdns_rdata_to_str() function is
printable US-ASCII, so this *should* be a safe sequence of conversions.