Paper over the differences between Python 2.7 and Python 3.4+ for name conversions
domain_to_str() must return a 'str' object. On Python 2, this is easy,
because 'str' and 'bytes' are the same type, but on Python 3, we must
explicitly decode into 'str'. Moreover, we need to avoid unnecessarily
calling .decode() on a 'str' object in Python 2, since this returns a
'unicode' object. So we cannot use the same code path for Python 2+3.

str_to_name(), despite its name, should accept either a 'str' or a
'bytes' object on Python 3, to make it possible to pass the result of
.encode()'ing a string, which returns a 'bytes' object, e.g., when
explicitly performing IDNA conversions. However, we need to avoid
unnecessarily calling .encode() on a 'bytes' object, because this method
does not exist on Python 3. So we cannot use the same code path for
Python 2+3.

The end result is that domain_to_str() and str_to_name() have more or
less similar behavior on Python 2 and Python 3, though IDNA conversions
are a bit more awkward than they should be:

    Python 2.7.9 (default, Dec 11 2014, 08:58:12)
    [GCC 4.9.2] on linux2
    Type "help", "copyright", "credits" or "license" for more information.
    >>> import wdns
    >>> wdns.str_to_name('www.example.com')
    '\x03www\x07example\x03com\x00'
    >>> wdns.domain_to_str(_)
    'www.example.com.'
    >>> u'😠.tld'.encode('idna')
    'xn--b38h.tld'
    >>> wdns.str_to_name(_)
    '\x08xn--b38h\x03tld\x00'
    >>> wdns.domain_to_str(_)
    'xn--b38h.tld.'
    >>> _.encode('ascii').decode('idna')
    u'\U0001f620.tld.'
    >>>

    Python 3.4.2 (default, Dec 27 2014, 13:16:08)
    [GCC 4.9.2] on linux
    Type "help", "copyright", "credits" or "license" for more information.
    >>> import wdns
    >>> wdns.str_to_name('www.example.com')
    b'\x03www\x07example\x03com\x00'
    >>> wdns.domain_to_str(_)
    'www.example.com.'
    >>> '😠.tld'.encode('idna')
    b'xn--b38h.tld'
    >>> wdns.str_to_name(_)
    b'\x08xn--b38h\x03tld\x00'
    >>> wdns.domain_to_str(_)
    'xn--b38h.tld.'
    >>> _.encode('ascii').decode('idna')
    '😠.tld.'
    >>>